Add type hints and mypy config
