+++ kart/tabular/v2.py
@@ -1,3 +1,4 @@
+from typing import ClassVar
 from kart.meta_items import MetaItemDefinition, MetaItemFileType
 from kart.tabular.v3 import TableV3
 from kart.tabular.v3_paths import PathEncoder
@@ -8,7 +9,7 @@ class TableV2(TableV3):
 
     DATASET_DIRNAME = ".sno-dataset"  # Old name for V2 datasets.
 
-    META_ITEMS = TableV3.META_ITEMS + (
+    META_ITEMS: ClassVar[tuple[MetaItemDefinition, ...]] = TableV3.META_ITEMS + (
         MetaItemDefinition("metadata/dataset.json", MetaItemFileType.JSON),
     )
