+++ mypy.ini
@@ -1,0 +1,34 @@
+[mypy]
+# when checking specific files (pre-commit), it's annoying when mypy reports a ton of errors in other files...
+# NOTE: this seems to be the default when run on Linux but not on MacOS? (Undocumented though)
+follow_imports = silent
+files = kart/
+python_version = 3.11
+
+[mypy-pysqlite3.*]
+ignore_missing_imports = true
+
+[mypy-osgeo.*]
+ignore_missing_imports = true
+
+[mypy-ipdb.*]
+ignore_missing_imports = true
+
+[mypy-shellingham.*]
+ignore_missing_imports = true
+
+[mypy-reflink.*]
+ignore_missing_imports = true
+
+[mypy-boto3.*]
+ignore_missing_imports = true
+
+[mypy-botocore.*]
+ignore_missing_imports = true
+
+[mypy-pyodbc.*]
+ignore_missing_imports = true
+
+[mypy-pygit2.*]
+# TODO: remove this when upgrading to pygit2 1.16+, which has builtin type hints
+ignore_missing_imports = true
