+++ requirements/dev.txt
@@ -1,37 +1,50 @@
 # This file was autogenerated by uv via the following command:
-#    uv pip compile dev.in -o dev.txt
+#    uv pip compile requirements/dev.in -o requirements/dev.txt
 appnope==0.1.3
     # via
-    #   -r dev.in
+    #   -r requirements/dev.in
     #   ipython
 asttokens==2.2.1
     # via stack-data
+attrs==22.1.0
+    # via
+    #   -c requirements/requirements.txt
+    #   -c requirements/test.txt
+    #   referencing
 backcall==0.2.0
     # via ipython
 colorama==0.4.6
     # via
-    #   -c docs.txt
-    #   -c test.txt
-    #   -r dev.in
+    #   -c requirements/docs.txt
+    #   -c requirements/test.txt
+    #   -r requirements/dev.in
 decorator==5.1.1
     # via
     #   ipdb
     #   ipython
 executing==1.2.0
     # via stack-data
+idna==3.7
+    # via
+    #   -c requirements/docs.txt
+    #   yarl
 ipdb==0.13.11
-    # via -r dev.in
+    # via -r requirements/dev.in
 ipython==8.10.0
     # via
-    #   -r dev.in
+    #   -r requirements/dev.in
     #   ipdb
 jedi==0.18.2
     # via ipython
 matplotlib-inline==0.1.6
     # via ipython
+msgpack-types==0.3.0
+    # via -r requirements/dev.in
+multidict==6.1.0
+    # via yarl
 mypy==1.13.0
     # via
-    #   -r dev.in
+    #   -r requirements/dev.in
     #   sqlalchemy-stubs
 mypy-extensions==1.0.0
     # via mypy
@@ -39,52 +52,60 @@ parso==0.8.3
     # via jedi
 pexpect==4.8.0
     # via
-    #   -r dev.in
+    #   -r requirements/dev.in
     #   ipython
 pickleshare==0.7.5
     # via ipython
 prompt-toolkit==3.0.36
     # via ipython
+propcache==0.2.0
+    # via yarl
 ptyprocess==0.7.0
     # via pexpect
 pure-eval==0.2.2
     # via stack-data
 pygments==2.13.0
     # via
-    #   -c docs.txt
-    #   -c requirements.txt
+    #   -c requirements/docs.txt
+    #   -c requirements/requirements.txt
     #   ipython
+pyrsistent==0.19.2
+    # via
+    #   -c requirements/requirements.txt
+    #   referencing
+referencing==0.8.11
+    # via types-jsonschema
 six==1.16.0
     # via
-    #   -c docs.txt
-    #   -c requirements.txt
-    #   -c test.txt
+    #   -c requirements/docs.txt
+    #   -c requirements/requirements.txt
+    #   -c requirements/test.txt
     #   asttokens
 sqlalchemy-stubs==0.4
-    # via -r dev.in
+    # via -r requirements/dev.in
 stack-data==0.6.2
     # via ipython
 traitlets==5.7.1
     # via
     #   ipython
     #   matplotlib-inline
-types-cffi==1.16.0.20240331
-    # via types-pygit2
 types-docutils==0.21.0.20241005
     # via types-pygments
-types-pygit2==1.14.0.20240317
-    # via -r dev.in
+types-jsonschema==4.23.0.20240813
+    # via -r requirements/dev.in
+types-psycopg2==2.9.21.20241019
+    # via -r requirements/dev.in
 types-pygments==2.13.0
-    # via -r dev.in
+    # via -r requirements/dev.in
 types-setuptools==75.3.0.20241107
-    # via
-    #   types-cffi
-    #   types-pygments
+    # via types-pygments
 types-tqdm==4.66.0.20240417
-    # via -r dev.in
+    # via -r requirements/dev.in
 typing-extensions==4.12.2
     # via
     #   mypy
     #   sqlalchemy-stubs
 wcwidth==0.2.5
     # via prompt-toolkit
+yarl==1.17.1
+    # via referencing
